refactor(procmgr): split Windows platform into focused modules
Mechanical split of platform/windows.rs into child_env, console,
job_object, process, runtime_user, and wide. No spawn or behavior change.
